updated payment constraint and skip bot issue solved

+++ Hippo/Model/HippoChannel.swift
@@ -364,7 +364,7 @@ class HippoChannel {
         params["grouping_tags"] = groupingTags
         
         if let skipBot = HippoProperty.current.skipBot {
-            params["skip_bot"] = skipBot
+            params["skip_bot"] = skipBot.intValue()
             params["skip_bot_reason"] = HippoProperty.current.skipBotReason ?? ""
         }
